TASK-17 & TASK-18: Jetpack Compose navigation & conditional navigation on startup
