+++ mobile/android/src/main/java/ch/levelup/kaesseli/android/navigation/Screen.kt
@@ -1,0 +1,9 @@
+package ch.levelup.kaesseli.android.navigation
+
+sealed class Screen(val route: String) {
+    object MainScreen: Screen("main")
+    object BalanceScreen: Screen("balance")
+    object OverviewScreen: Screen("overview")
+    object RegisterScreen: Screen("register")
+    object CodeSubmissionScreen: Screen("code_submission")
+}
