--- mobile/android/src/main/res/navigation/main_graph.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/main_graph"
-    app:startDestination="@id/mainFragment">
-
-    <fragment
-        android:id="@+id/mainFragment"
-        android:name="ch.levelup.kaesseli.android.MainFragment"
-        android:label="MainFragment"/>
-    <fragment
-        android:id="@+id/codeSubmission"
-        android:name="ch.levelup.kaesseli.android.CodeSubmission"
-        android:label="CodeSubmission">
-        <action
-            android:id="@+id/action_codeSubmission_to_overview"
-            app:destination="@id/overview"
-            app:enterAnim="@android:anim/fade_in"
-            app:exitAnim="@android:anim/fade_out" />
-
-        <action
-            android:id="@+id/action_codeSubmission_to_register"
-            app:destination="@id/register"
-            app:enterAnim="@android:anim/fade_in"
-            app:exitAnim="@android:anim/fade_out" />
-    </fragment>
-    <fragment
-        android:id="@+id/register"
-        android:name="ch.levelup.kaesseli.android.Register"
-        android:label="Register">
-        <action
-            android:id="@+id/action_register_to_overview"
-            app:destination="@id/overview"
-            app:enterAnim="@android:anim/fade_in"
-            app:exitAnim="@android:anim/fade_out" />
-    </fragment>
-    <fragment
-        android:id="@+id/overview"
-        android:name="ch.levelup.kaesseli.android.Overview"
-        android:label="Overview">
-        <action
-            android:id="@+id/action_overview_to_balance"
-            app:destination="@id/balance"
-            app:enterAnim="@android:anim/fade_in"
-            app:exitAnim="@android:anim/fade_out" />
-    </fragment>
-    <fragment
-        android:id="@+id/balance"
-        android:name="ch.levelup.kaesseli.android.Balance"
-        android:label="Balance" />
-</navigation>
